+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+<a name="3.4.0"></a>
+## [3.4.0](https://github.com/gabyvs/ng2-ue-utils/compare/3.3.1...3.4.0) (2017-02-27)
+
+### Features
+
+* **ClientObserver:** New reset method.
+* **ModalBase:** More flexible modal. Some classes included to replicate styles of ue SPAs.
+
 <a name="3.3.1"></a>
 ## [3.3.1](https://github.com/gabyvs/ng2-ue-utils/compare/3.3.0...3.3.1) (2017-02-22)